Added typechecker versions of `Expr` and `Const`
These versions of `Expr`, `Const` and `Univ` all have pointers
connecting their nodes, instead of CIDs, which allows for faster
retrieval of children nodes, which is a must for a fast reducer.
There are `Hash` fields in `Expr` and `Const`, which are just numbers
in 1-1 correspondence with Anon CIDs. The only reason I chose `Nat`
instead of the CID itself is that number equality is faster.