Adding typechecker to Lean Yatima (#64)
* Added typechecker universe reduction, instantiation and equality

* Added typechecker versions of `Expr` and `Const`

These versions of `Expr`, `Const` and `Univ` all have pointers
connecting their nodes, instead of CIDs, which allows for faster
retrieval of children nodes, which is a must for a fast reducer.
There are `Hash` fields in `Expr` and `Const`, which are just numbers
in 1-1 correspondence with Anon CIDs. The only reason I chose `Nat`
instead of the CID itself is that number equality is faster.

* Added `Value` (i.e. reduced expressions)

* Added evaluation of expressions, missing `proj` case

* Equality of expressions added, with proof irrelevance

* Added `check` and `infer`

* Adding IPLD Expr -> Typechecker.Expr translations

* IPLD Univ to typechecker Univ

* Added conversion of constants

Almost finished IPLD conversion. The missing case is the function that
checks if an inductive is unit-like. Also, we should probably consider
adding conversion cache.

* Added projections

* Fixed projection inference

* Added print functions for debugging

* WIP Refactoring of Const

* fill in TODOs, but there's a bug

* fix internal/external recursor detection but there are test failures

* bump Lean toolchain

* update

* update

* fix CID test

* put back `ExprGenerator.lean`

* add CI

* minor

* WIP fixing typechecker translator

* Finished conversion of IPLD into Typechecker

Have not tested yet; there might be some bugs.

* Fixed printing

* Fixed compilation of recursion rules

* Added translation cache + refactoring

* Some refactoring

* Some more refactoring

* Added conversion of store to typechecker expressions

* CLI update + refactoring

* formatting and dropping Graph.lean and its dependency functions

Co-authored-by: Rishikesh Vaishnav <rishhvaishnav@gmail.com>
Co-authored-by: Arthur Paulino <arthurleonardo.ap@gmail.com>