A refactoring which removes cyclic structures
Constants are much like definitions now. A `.constant` expression is
just a reference to a global environment of constants, which is an
array for efficiency. This removes the need of `Fix` since Lean does
not have internal recursion, only recursion at the top-level (`let
rec`  is actually a macro that lifts the definition to the top-level).
Recursion (mutual or not) is now simply a matter of referencing other
constants. I've removed `ConstCid -> Const` mapping. `Yatima.Store` is
now `Ipld.Store` and it is purely `IPLD`. On the other hand a new
"store" appears, an array of `Const` called `defns`, which is purely
used for typechecking.