Refactoring of the typechecker
`Expr` does not have `TypeInfo` anymore. Instead, `infer` and `check`
both return `TypedExpr`, which is an IR for evaluation and
compilation. Eventually, `TypedExpr` will annotate erased values and
such. `TypeInfo` has been changed into an inductive, since all the
flags are mutually exclusive. The `struct?` flag on `TypeInfo` has
been removed, but the structure index has been added to
`TypedExpr.proj` which is needed for equality of projections. On that
note, equality now takes suspensions instead of values, in order to
not evaluate proofs. Lastly, I've removed the `lam` and `letE` cases
in check, and will eventually add the `lam` case to infer.

The `TypedExpr` refactoring made visible a bug that wasn't
noticed. The expressions that we got from dereferencing constants were
not correctly populated with `TypeInfo`. We will have to, in the
process of typechecking, building a new const store, which derefences
to `TypedExpr`